+++ deckster/common/configs.py
@@ -18,11 +18,20 @@ dir_path = f"{str(Path.home())}/.config/deckster/"
 
 logger.debug(f"Config path set to: {dir_path}")
 
-# Static variable for lock plugin
-class counter:
+
+class _counter:
     count = 0
 
 def read_config(cfg, custom_config = None):
+    """Read the configuration file
+
+    Args:
+        cfg (string): the config to read (key)
+        custom_config (string, optional): Sepcify an alternate path. Defaults to None.
+
+    Returns:
+        string: value
+    """
     if custom_config == None:
         full_cfg = os.path.join(dir_path, "config.json")
     else:
@@ -42,6 +51,13 @@ def read_config(cfg, custom_config = None):
         return None
 
 def write_config(cfg, value, custom_config = None):
+    """Write value for a config
+
+    Args:
+        cfg (string): config to update
+        value (string): value to write
+        custom_config (string, optional): Provide an alternate path. Defaults to None.
+    """
     if custom_config == None:
         full_cfg = os.path.join(dir_path, "config.json")
     else:
@@ -56,7 +72,7 @@ def write_config(cfg, value, custom_config = None):
             logger.debug(f"Writing '{value}' to '{cfg}'")
             json.dump(data, cfgFile, indent=2)
 
-def fetch_templated_key(json_key):
+def _fetch_templated_keys(json_key):
     template_name = json_key["template"]
     keysdir = os.path.expanduser(read_config("keys_dir"))
     tmpl_files = [f for f in os.listdir(keysdir) if os.path.isfile(os.path.join(keysdir, f)) and (os.path.splitext(f)[1] in [".tmpl"])]
@@ -70,6 +86,11 @@ def fetch_templated_key(json_key):
     
 # Add lock here too
 def read_keys():
+    """Read all keys
+
+    Returns:
+        list: A list containing all keys
+    """
     logger.debug(f"Reading keys...")
     templist = []
     keysdir = os.path.expanduser(read_config("keys_dir"))
@@ -93,17 +114,35 @@ def read_keys():
             logger.error(f"Error parsing {os.path.join(keysdir, files)}: {e}")   
     for x in json_keys:
         if "template" in x:
-            x = fetch_templated_key(x)
+            x = _fetch_templated_keys(x)
         templist.append(Key(x))
     return templist
 
 def find_key(key, page, key_list):
+    """Find specific key on a page
+
+    Args:
+        key (integer): The key number
+        page (integer): The page to search the key
+        key_list (list): A list of keys
+
+    Returns:
+        Key: A found key
+    """
     logger.debug(f"Finding key {key} on page {page}.")
     for x in key_list:
         if x.key == key and x.page == page:
             return x
 
 def max_page(key_list):
+    """Find the last page
+
+    Args:
+        key_list (list): List of keys
+
+    Returns:
+        integer: Page number
+    """
     high = 0
     for x in key_list:
         if x.page > high:
@@ -111,6 +150,14 @@ def max_page(key_list):
     return high
 
 def write_key_config(key, page, cfg, value):
+    """Write value for a given config for a given key.
+
+    Args:
+        key (Key): A key
+        page (integer): Page number
+        cfg (string): config to write
+        value (string): value to write
+    """
     keysdir = os.path.expanduser(read_config("keys_dir"))
     logger.debug(f"{keysdir} is the keys directory.")
     keyfiles = [f for f in os.listdir(keysdir) if os.path.isfile(os.path.join(keysdir, f))]
@@ -154,6 +201,15 @@ def write_key_config(key, page, cfg, value):
         logger.error(f"Could not find a match for key {key} on page {page}.")
 
 def empty_set(key_count, page):
+    """Creates empty keys
+
+    Args:
+        key_count (integer): Number of keys to generate
+        page (integer): The page to assign to generated keys
+
+    Returns:
+        list: A list of empty keys
+    """
     ks = []
     for k in range(key_count):
         ks.append(fake_key(k, "empty", page))